Centralizes shared session models in ag-session
- Exposes orchestration, personality, project, clarification, review, setting, and transcript models plus pure helpers through the shared crate.
- Preserves Agentty compatibility re-exports while keeping app-local input state in `domain`.
- Aligns crate guidance and architecture documentation with the new ownership boundaries.
