+++ crates/AGENTS.md
@@ -18,8 +18,9 @@ Contains the workspace member crates.
 - `crates/ag-protocol/` holds structured agent response contracts, protocol parsing,
   schema generation, and transport-neutral turn prompt payloads shared across frontends
   and agent adapters.
-- `crates/ag-session/` holds shared session identity, lifecycle and transcript models,
-  complete session aggregates, and the frontend-neutral programmatic lifecycle API.
+- `crates/ag-session/` holds shared session identity, lifecycle, orchestration, project,
+  personality, review, setting, clarification, and transcript models plus the complete
+  session aggregates and frontend-neutral programmatic lifecycle API.
 - `crates/ag-tui-text/` holds shared Ratatui text rendering helpers for markdown,
   mermaid diagrams, and terminal-width wrapping/truncation.
 - `crates/ag-xtask/` holds workspace maintenance utilities, including migration checks.
