+++ crates/agentty/src/AGENTS.md
@@ -4,8 +4,9 @@
 
 - Avoid near-identical local variable names in the same function (for example, `gitdir`
   and `git_dir`). Use one clear naming style with distinct, descriptive names.
-- Persisted keys in `setting` and `project_setting` must come from `domain/setting.rs`
-  `SettingName`; do not introduce ad hoc string keys or legacy aliases.
+- Persisted keys in `setting` and `project_setting` must come from
+  `ag_session::SettingName`, re-exported by `domain/setting.rs`; do not introduce ad hoc
+  string keys or legacy aliases.
 - Prefer `module.rs` plus `module/` for nested modules. Avoid `mod.rs` module roots.
 - Session status flow:
   - `SessionStatus::can_transition_to()` in `crates/ag-session/src/model.rs` is the
@@ -48,8 +49,8 @@ When changing architecture-level behavior under `src/`, update:
 ## Major Areas
 
 - `app.rs` and `app/` own orchestration and workflow state.
-- `domain.rs` and `domain/` own Agentty-specific business entities and re-export shared
-  session models from `ag-session`.
+- `domain.rs` and `domain/` own Agentty-specific business entities and compatibility
+  re-exports for shared models from `ag-session`.
 - `infra.rs` and `infra/` own external integrations and persistence.
 - `runtime.rs` and `runtime/` own terminal lifecycle and event dispatch.
 - `ui.rs` and `ui/` own rendering, layout, and interaction widgets.
