+++ crates/agentty/src/domain/AGENTS.md
@@ -9,11 +9,13 @@ Pure business logic and domain entities, decoupled from UI and infrastructure.
 - `agent.rs` defines provider kinds, models, and model metadata.
 - `session.rs` defines Agentty render/runtime session entities and sizing logic while
   re-exporting shared identity, status, and review-link models from `ag-session`.
-- `orchestration.rs` owns pure plan validation, child-status mapping, parallelism, and
-  roll-up readiness policy; persistence and session-runtime adapters remain in `app/`.
+- `orchestration.rs`, `personality.rs`, `project.rs`, `review.rs`, `setting.rs`, and
+  `transcript_notice.rs` preserve Agentty's compatibility paths for models owned by
+  `ag-session`.
+- `question.rs` owns app-local input progress while re-exporting shared clarification
+  models and option-selection policy from `ag-session`.
 - `session_message.rs` re-exports durable transcript models from `ag-session`.
-- `project.rs`, `setting.rs`, `permission.rs`, and `input.rs` define shared application
-  concepts.
+- `permission.rs` and `input.rs` define shared application concepts.
 
 ## Docs
